Commit current changes
